% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAOP.RAA.2024-package.R
\docType{package}
\name{CAOP.RAA.2024-package}
\alias{CAOP.RAA.2024}
\alias{CAOP.RAA.2024-package}
\title{CAOP.RAA.2024: Official Administrative Map of the Azores (CAOP 2024)}
\description{
Provides the official administrative boundaries of the Azores (Região Autónoma dos Açores (RAA)) as defined in the 2024 edition of the Carta Administrativa Oficial de Portugal (CAOP), published by the Direção-Geral do Território (DGT). The package includes convenience functions to import these boundaries as 'sf' objects for spatial analysis in R. Source: \url{https://geo2.dgterritorio.gov.pt/caop/CAOP_RAA_2024-gpkg.zip}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patterninstitute/CAOP.RAA.2024}
  \item \url{https://www.pattern.institute/CAOP.RAA.2024/}
  \item Report bugs at \url{https://github.com/patterninstitute/CAOP.RAA.2024/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
