% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islands.R
\name{islands}
\alias{islands}
\title{Azorean islands}
\usage{
islands()
}
\value{
A \link[tibble:tibble]{tibble} with columns:

\describe{
\item{\code{group}}{Island group.}
\item{\code{is_triangle}}{Due to their proximity, Pico, Faial and São Jorge form
the so-called \emph{Triangle} of the Central Group of islands.}
\item{\code{name}}{Island's name.}
}
}
\description{
A brief dataset of the nine Azorean islands.
}
\examples{
islands()

}
